requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
PyYAML==6.0.2
pydantic==2.9.2
openai>=1.40.0
load_dotenv